pyproject.toml
[project]
name = "streamlit_app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.3",
    "pandas-stubs~=2.3.3",
    "plotly>=6.6.0",
    "scipy>=1.17.1",
    "scipy-stubs~=1.17.1",
    "streamlit>=1.55.0",
    "streamlit-js-eval>=1.0.0",
]
